backend/requirements.txt
# FastAPI and Core Dependencies
fastapi[standard]==0.115.13
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0

# Database
sqlmodel==0.0.22
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# Background Tasks
celery==5.4.0
redis==5.2.1

# File Processing & OCR
python-magic==0.4.27
PyMuPDF==1.25.2
pdfplumber==0.11.4
python-docx==1.1.2
Pillow==11.0.0
pytesseract==0.3.13
pdf2image==1.17.0

# S3/MinIO Storage
boto3==1.35.95
minio==7.2.11

# AI/LLM Integration
openai==1.59.3
anthropic==0.42.0
crewai==0.152.0
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.2.14

# HTTP Client
httpx==0.28.1
aiofiles==24.1.0

# Validation & Parsing
email-validator==2.2.0
phonenumbers==8.13.52

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.14.0

# Monitoring & Logging
structlog==24.5.0
prometheus-client==0.21.1

# Utilities
python-dotenv==1.0.1
click==8.1.8
rich==13.9.4
